@@ -1,6 +1,6 @@
 [package]
 name = "lattice-qcd-rs"
-version = "0.2.1"
+version = "0.3.0"
 authors = ["Aliénore Bouttefeux <alienore.bouttefeux@gmail.com>"]
 edition = "2021"
 readme = "README.md"
@@ -10,39 +10,49 @@ description = "Lattice QCD simulation"
 keywords = ["QCD", "Lattice", "Monte-Carlo", "chromodynamics"]
 publish = true
 license = "MIT OR Apache-2.0"
-rust-version = "1.61"
+rust-version = "1.72"
+
 
 [lib]
 path = "src/lib.rs"
 crate-type = ["lib"]
 
+
 [features]
-serde-serialize = ["serde", "nalgebra/serde-serialize"]
-no-overflow-test= []
-default = ["serde-serialize"]
+serde-serialize = ["dep:serde", "nalgebra/serde-serialize", "utils-lib/serde"]
+overflow-test = []
+# experimental = []
+default = ["serde-serialize", "overflow-test"]
+
 
 [workspace]
 members = ["procedural_macro"]
 
+
 [dependencies]
-nalgebra = { version = "0.32.2", features = ["serde-serialize"] }
+nalgebra = { version = "0.32", features = ["serde-serialize"] }
 approx = "0.5.1"
-num-traits = "0.2.15"
+num-traits = "0.2.17"
 rand = "0.8.5"
 rand_distr = "0.4.3"
-crossbeam = "0.8.1"
-rayon = "1.5.3"
+crossbeam = "0.8.4"
+rayon = "1.8"
 serde = { version = "1.0", features = ["derive"], optional = true }
-lattice_qcd_rs-procedural_macro = {path = "procedural_macro", version = "0.2.1"}
+lattice_qcd_rs-procedural_macro = { path = "procedural_macro", version = "0.3.0" }
+utils-lib = { version = "0.1.0", git = "https://github.com/ABouttefeux/utils-lib", branch = "main", features = [
+    "serde",
+] }
 
 [dev-dependencies]
 criterion = "0.5.1"
-version-sync = "0.9.4"
+version-sync = "0.9.5"
+
 
 [[bench]]
 name = "bench"
 path = "benches/bench.rs"
 harness = false
+
 
 [profile.test]
 opt-level = 2 # in order to have resonable test time
