@@ -1,17 +1,24 @@
-extern crate nalgebra as na;
-
-use std::{collections::HashMap, f64, vec::Vec};
+use std::{collections::HashMap, f64};
 
 use criterion::{criterion_group, criterion_main, BatchSize, BenchmarkId, Criterion, Throughput};
 use lattice_qcd_rs::{
-    field::*, integrator::*, simulation::*, su3::su3_exp_i, su3::su3_exp_r, Complex, Real,
+    field::{EField, LinkMatrix, Su3Adjoint},
+    integrator::{SymplecticEuler, SymplecticEulerRayon},
+    simulation::{
+        HybridMonteCarloDiagnostic, LatticeState, LatticeStateDefault, LatticeStateEFSyncDefault,
+        MetropolisHastingsSweep, SimulationStateSynchronous,
+    },
+    su3::{su3_exp_i, su3_exp_r},
+    Complex, Real,
 };
+use rand::rngs::ThreadRng;
+use rand_distr::Distribution;
 use rayon::prelude::*;
 
 fn bench_simulation_creation_determinist(
     size: usize,
-    rng: &mut rand::rngs::ThreadRng,
-    d: &impl rand_distr::Distribution<Real>,
+    rng: &mut ThreadRng,
+    d: &impl Distribution<Real>,
 ) {
     let _simulation = LatticeStateEFSyncDefault::<LatticeStateDefault<4>, 4>::new_determinist(
         1_f64, 1_f64, size, rng, d,
@@ -21,7 +28,7 @@ fn bench_simulation_creation_determinist(
 
 fn bench_simulation_creation_threaded<D>(size: usize, d: &D, number_of_thread: usize)
 where
-    D: rand_distr::Distribution<Real> + Sync,
+    D: Distribution<Real> + Sync,
 {
     let _simulation = LatticeStateEFSyncDefault::<LatticeStateDefault<4>, 4>::new_random_threaded(
         1_f64,
@@ -33,30 +40,26 @@ where
     .unwrap();
 }
 
-fn matrix_exp_old(rng: &mut rand::rngs::ThreadRng, d: &impl rand_distr::Distribution<Real>) {
-    let _m = (Su3Adjoint::random(rng, d).to_matrix() * na::Complex::<Real>::i()).exp();
+fn matrix_exp_old(rng: &mut ThreadRng, d: &impl Distribution<Real>) {
+    let _m = (Su3Adjoint::random(rng, d).to_matrix() * nalgebra::Complex::<Real>::i()).exp();
 }
 
-fn matrix_exp_i(rng: &mut rand::rngs::ThreadRng, d: &impl rand_distr::Distribution<Real>) {
+fn matrix_exp_i(rng: &mut ThreadRng, d: &impl Distribution<Real>) {
     let _m = su3_exp_i(Su3Adjoint::random(rng, d));
 }
 
-fn matrix_exp_r(rng: &mut rand::rngs::ThreadRng, d: &impl rand_distr::Distribution<Real>) {
+fn matrix_exp_r(rng: &mut ThreadRng, d: &impl Distribution<Real>) {
     let _m = su3_exp_r(Su3Adjoint::random(rng, d));
 }
 
-fn create_su3_adj(rng: &mut rand::rngs::ThreadRng, d: &impl rand_distr::Distribution<Real>) {
+fn create_su3_adj(rng: &mut ThreadRng, d: &impl Distribution<Real>) {
     let _m = Su3Adjoint::random(rng, d);
 }
 
-fn create_vec(
-    rng: &mut rand::rngs::ThreadRng,
-    d: &impl rand_distr::Distribution<Real>,
-    size: usize,
-) {
+fn create_vec(rng: &mut ThreadRng, d: &impl Distribution<Real>, size: usize) {
     let mut data = Vec::with_capacity(size);
     for _ in 0..size {
-        data.push(na::Vector4::new(
+        data.push(nalgebra::Vector4::new(
             Su3Adjoint::random(rng, d),
             Su3Adjoint::random(rng, d),
             Su3Adjoint::random(rng, d),
@@ -65,16 +68,12 @@ fn create_vec(
     }
 }
 
-fn create_hash_map(
-    rng: &mut rand::rngs::ThreadRng,
-    d: &impl rand_distr::Distribution<Real>,
-    size: usize,
-) {
+fn create_hash_map(rng: &mut ThreadRng, d: &impl Distribution<Real>, size: usize) {
     let mut data = HashMap::with_capacity(size);
     for i in 0..size {
         data.insert(
             i,
-            na::Vector4::new(
+            nalgebra::Vector4::new(
                 Su3Adjoint::random(rng, d),
                 Su3Adjoint::random(rng, d),
                 Su3Adjoint::random(rng, d),
@@ -106,11 +105,11 @@ fn criterion_benchmark(c: &mut Criterion) {
     let mut groupe_creation_determinist = c.benchmark_group("Sim creation determinist");
     groupe_creation_determinist.sample_size(10);
     let array_size: [usize; 6] = [2, 4, 8, 10, 20, 30];
-    for el in array_size.iter() {
+    for el in array_size {
         groupe_creation_determinist.throughput(Throughput::Elements((el.pow(4)) as u64));
         groupe_creation_determinist.bench_with_input(
             BenchmarkId::new("size", el.pow(4)),
-            el,
+            &el,
             |b, i| b.iter(|| bench_simulation_creation_determinist(*i, &mut rng, &d)),
         );
     }
@@ -121,9 +120,9 @@ fn criterion_benchmark(c: &mut Criterion) {
     groupe_creation_threaded.sample_size(50);
     groupe_creation_threaded.throughput(Throughput::Elements(gp_size.pow(4) as u64));
     let thread_count: [usize; 5] = [1, 2, 4, 6, 8];
-    for n in thread_count.iter() {
-        groupe_creation_threaded.bench_with_input(BenchmarkId::new("thread", n), n, |b, i| {
-            b.iter(|| bench_simulation_creation_threaded(gp_size, &d, *i))
+    for n in thread_count {
+        groupe_creation_threaded.bench_with_input(BenchmarkId::new("thread", n), &n, |b, i| {
+            b.iter(|| bench_simulation_creation_threaded(gp_size, &d, *i));
         });
     }
     groupe_creation_threaded.finish();
@@ -131,26 +130,26 @@ fn criterion_benchmark(c: &mut Criterion) {
     let mut groupe_matrix = c.benchmark_group("matrix exp");
     groupe_matrix.sample_size(2_000);
     groupe_matrix.bench_function("matrix exp old", |b| {
-        b.iter(|| matrix_exp_old(&mut rng, &d))
+        b.iter(|| matrix_exp_old(&mut rng, &d));
     });
     groupe_matrix.bench_function("matrix exp new i ", |b| {
-        b.iter(|| matrix_exp_i(&mut rng, &d))
+        b.iter(|| matrix_exp_i(&mut rng, &d));
     });
     groupe_matrix.bench_function("matrix exp new r", |b| {
-        b.iter(|| matrix_exp_r(&mut rng, &d))
+        b.iter(|| matrix_exp_r(&mut rng, &d));
     });
     groupe_matrix.finish();
 
     let mut groupe_sim = c.benchmark_group("simulations");
     groupe_sim.sample_size(10);
     let thread_count: [usize; 5] = [1, 2, 4, 6, 8];
-    for n in thread_count.iter() {
+    for n in thread_count {
         let mut sim = LatticeStateEFSyncDefault::<LatticeStateDefault<4>, 4>::new_random_threaded(
             1_f64, 1_f64, 5, &d, 4,
         )
         .unwrap();
-        groupe_sim.bench_with_input(BenchmarkId::new("thread", n), n, |b, i| {
-            b.iter(|| simulate_euler(&mut sim, *i))
+        groupe_sim.bench_with_input(BenchmarkId::new("thread", n), &n, |b, i| {
+            b.iter(|| simulate_euler(&mut sim, *i));
         });
     }
 
@@ -159,7 +158,7 @@ fn criterion_benchmark(c: &mut Criterion) {
     )
     .unwrap();
     groupe_sim.bench_function("simulate(20) rayon", |b| {
-        b.iter(|| simulate_euler_rayon(&mut sim))
+        b.iter(|| simulate_euler_rayon(&mut sim));
     });
     groupe_sim.finish();
 
@@ -173,7 +172,7 @@ fn criterion_benchmark(c: &mut Criterion) {
             || LatticeStateDefault::<3>::new_determinist(1000_f64, 2_f64, 20, &mut rng).unwrap(),
             |state_in| state_in.monte_carlo_step(&mut mc),
             BatchSize::LargeInput,
-        )
+        );
     });
 
     let mut mch =
@@ -184,7 +183,7 @@ fn criterion_benchmark(c: &mut Criterion) {
             || LatticeStateDefault::<3>::new_determinist(1000_f64, 2_f64, 20, &mut rng).unwrap(),
             |state_in| state_in.monte_carlo_step(&mut mch),
             BatchSize::LargeInput,
-        )
+        );
     });
     groupe_mc.finish();
 
@@ -192,29 +191,29 @@ fn criterion_benchmark(c: &mut Criterion) {
     groupe_gauss_proj.sample_size(10);
 
     let array_size: [usize; 4] = [2, 4, 8, 10];
-    for n in array_size.iter() {
-        groupe_gauss_proj.bench_with_input(BenchmarkId::new("size", n), n, |b, i| {
+    for n in array_size {
+        groupe_gauss_proj.bench_with_input(BenchmarkId::new("size", n), &n, |b, i| {
             b.iter(|| {
                 let state =
                     LatticeStateDefault::<4>::new_determinist(1_f64, 1_f64, *i, &mut rng).unwrap();
                 let d = rand_distr::Normal::new(0.0, 0.5_f64).unwrap();
                 let e_field = EField::new_determinist(state.lattice(), &mut rng, &d);
-                e_field.project_to_gauss(state.link_matrix(), state.lattice());
-            })
+                let _ = e_field.project_to_gauss(state.link_matrix(), state.lattice());
+            });
         });
     }
     groupe_gauss_proj.finish();
 
     let mut groupe_clone = c.benchmark_group("Clone");
     groupe_clone.sample_size(100);
-    let link_m = LinkMatrix::new(vec![na::Matrix3::<Complex>::identity(); 10_000]);
+    let link_m = LinkMatrix::new(vec![nalgebra::Matrix3::<Complex>::identity(); 10_000]);
     groupe_clone.bench_function("rayon clone", |b| {
-        b.iter(|| LinkMatrix::new(link_m.data().par_iter().copied().collect()))
+        b.iter(|| LinkMatrix::new(link_m.data().par_iter().copied().collect()));
     });
 
     groupe_clone.bench_function("clone", |b| b.iter(|| link_m.clone()));
 
-    groupe_clone.finish()
+    groupe_clone.finish();
 }
 
 fn benchmark_base(c: &mut Criterion) {
@@ -223,22 +222,26 @@ fn benchmark_base(c: &mut Criterion) {
 
     let mut groupe_collection = c.benchmark_group("std collections");
     groupe_collection.sample_size(10);
-    let size: [usize; 5] = [100, 1_000, 10_000, 100_000, 1_000_000];
-    for s in size.iter() {
-        groupe_collection.throughput(Throughput::Elements(*s as u64));
-        groupe_collection.bench_with_input(BenchmarkId::new("vec size", s), s, |b, i| {
-            b.iter(|| create_vec(&mut rng, &d, *i))
+    let sizes: [usize; 5] = [100, 1_000, 10_000, 100_000, 1_000_000];
+    for size in sizes {
+        groupe_collection.throughput(Throughput::Elements(size as u64));
+        groupe_collection.bench_with_input(BenchmarkId::new("vec size", size), &size, |b, i| {
+            b.iter(|| create_vec(&mut rng, &d, *i));
         });
-        groupe_collection.bench_with_input(BenchmarkId::new("hash_map size", s), s, |b, i| {
-            b.iter(|| create_hash_map(&mut rng, &d, *i))
-        });
+        groupe_collection.bench_with_input(
+            BenchmarkId::new("hash_map size", size),
+            &size,
+            |b, i| {
+                b.iter(|| create_hash_map(&mut rng, &d, *i));
+            },
+        );
     }
     groupe_collection.finish();
 
     let mut groupe_random = c.benchmark_group("random gen");
     groupe_random.sample_size(2_000);
     groupe_random.bench_function("creation time random SU3 adj", |b| {
-        b.iter(|| create_su3_adj(&mut rng, &d))
+        b.iter(|| create_su3_adj(&mut rng, &d));
     });
     groupe_random.finish();
 }
