@@ -1,23 +1,21 @@
 //! Basic symplectic Euler integrator using [`Rayon`](https://docs.rs/rayon/1.5.1/rayon/).
 
-use std::vec::Vec;
+use std::fmt::{self, Display};
 
-use na::SVector;
+use nalgebra::SVector;
 #[cfg(feature = "serde-serialize")]
 use serde::{Deserialize, Serialize};
 
-use super::{
-    super::{
-        field::{EField, LinkMatrix, Su3Adjoint},
-        lattice::LatticeCyclic,
-        simulation::{
-            LatticeState, LatticeStateWithEField, LatticeStateWithEFieldNew, SimulationStateLeap,
-            SimulationStateSynchronous,
-        },
-        thread::run_pool_parallel_rayon,
-        Real,
+use super::{integrate_efield, integrate_link, CMatrix3, SymplecticIntegrator};
+use crate::{
+    field::{EField, LinkMatrix, Su3Adjoint},
+    lattice::LatticeCyclic,
+    simulation::{
+        LatticeState, LatticeStateWithEField, LatticeStateWithEFieldNew, SimulationStateLeap,
+        SimulationStateSynchronous,
     },
-    integrate_efield, integrate_link, CMatrix3, SymplecticIntegrator,
+    thread::run_pool_parallel_rayon,
+    Real,
 };
 
 /// Basic symplectic Euler integrator using Rayon.
@@ -52,16 +50,20 @@ use super::{
 pub struct SymplecticEulerRayon;
 
 impl SymplecticEulerRayon {
-    /// Create a new SymplecticEulerRayon
+    /// Create a new [`SymplecticEulerRayon`]
+    #[must_use]
+    #[inline]
     pub const fn new() -> Self {
-        Self {}
+        Self
     }
 
     /// Get all the integrated links
     /// # Panics
-    /// panics if the lattice has fewer link than link_matrix has or it has fewer point than e_field has.
-    /// In debug panic if the lattice has not the same number link as link_matrix
-    /// or not the same number of point as e_field.
+    /// panics if the lattice has fewer link than `link_matrix` has or it has fewer point than `e_field` has.
+    /// In debug panic if the lattice has not the same number link as `link_matrix`
+    /// or not the same number of point as `e_field`.
+    #[must_use]
+    #[inline]
     fn link_matrix_integrate<State, const D: usize>(
         link_matrix: &LinkMatrix,
         e_field: &EField<D>,
@@ -71,6 +73,7 @@ impl SymplecticEulerRayon {
     where
         State: LatticeStateWithEField<D>,
     {
+        // TODO improve perf
         run_pool_parallel_rayon(
             lattice.get_links(),
             &(link_matrix, e_field, lattice),
@@ -82,9 +85,11 @@ impl SymplecticEulerRayon {
 
     /// Get all the integrated e field
     /// # Panics
-    /// panics if the lattice has fewer link than link_matrix has or it has fewer point than e_field has.
-    /// In debug panic if the lattice has not the same number link as link_matrix
-    /// or not the same number of point as e_field.
+    /// panics if the lattice has fewer link than `link_matrix` has or it has fewer point than `e_field` has.
+    /// In debug panic if the lattice has not the same number link as `link_matrix`
+    /// or not the same number of point as `e_field`.
+    #[must_use]
+    #[inline]
     fn e_field_integrate<State, const D: usize>(
         link_matrix: &LinkMatrix,
         e_field: &EField<D>,
@@ -94,6 +99,7 @@ impl SymplecticEulerRayon {
     where
         State: LatticeStateWithEField<D>,
     {
+        // TODO improve perf
         run_pool_parallel_rayon(
             lattice.get_points(),
             &(link_matrix, e_field, lattice),
@@ -106,13 +112,15 @@ impl SymplecticEulerRayon {
 
 impl Default for SymplecticEulerRayon {
     /// Identical to [`SymplecticEulerRayon::new`].
+    #[inline]
     fn default() -> Self {
         Self::new()
     }
 }
 
-impl std::fmt::Display for SymplecticEulerRayon {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+impl Display for SymplecticEulerRayon {
+    #[inline]
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "Euler integrator using rayon")
     }
 }
@@ -124,6 +132,7 @@ where
 {
     type Error = State::Error;
 
+    #[inline]
     fn integrate_sync_sync(&self, l: &State, delta_t: Real) -> Result<State, Self::Error> {
         let link_matrix = Self::link_matrix_integrate::<State, D>(
             l.link_matrix(),
@@ -143,6 +152,7 @@ where
         )
     }
 
+    #[inline]
     fn integrate_leap_leap(
         &self,
         l: &SimulationStateLeap<State, D>,
@@ -170,6 +180,7 @@ where
         )
     }
 
+    #[inline]
     fn integrate_sync_leap(
         &self,
         l: &State,
@@ -192,6 +203,7 @@ where
         )
     }
 
+    #[inline]
     fn integrate_leap_sync(
         &self,
         l: &SimulationStateLeap<State, D>,
@@ -219,6 +231,7 @@ where
         )
     }
 
+    #[inline]
     fn integrate_symplectic(&self, l: &State, delta_t: Real) -> Result<State, Self::Error> {
         // override for optimization.
         // This remove a clone operation.
