@@ -41,24 +41,22 @@
 //! # }
 //! ```
 
-use rand_distr::Distribution;
+use std::fmt::{self, Display};
+
+use rand_distr::{Bernoulli, Distribution, Uniform};
 #[cfg(feature = "serde-serialize")]
 use serde::{Deserialize, Serialize};
 
-use super::{
-    super::{
-        super::{
-            error::Never,
-            field::LinkMatrix,
-            lattice::{
-                Direction, LatticeCyclic, LatticeElementToIndex, LatticeLink, LatticeLinkCanonical,
-                LatticePoint,
-            },
-            su3, Complex, Real,
-        },
-        state::{LatticeState, LatticeStateDefault, LatticeStateNew},
+use super::{delta_s_old_new_cmp, MonteCarlo, MonteCarloDefault};
+use crate::{
+    error::Never,
+    field::LinkMatrix,
+    lattice::{
+        Direction, LatticeCyclic, LatticeElementToIndex, LatticeLink, LatticeLinkCanonical,
+        LatticePoint,
     },
-    delta_s_old_new_cmp, MonteCarlo, MonteCarloDefault,
+    simulation::state::{LatticeState, LatticeStateDefault, LatticeStateNew},
+    su3, Complex, Real,
 };
 
 /// Metropolis Hastings algorithm. Very slow, use [`MetropolisHastingsDeltaDiagnostic`]
@@ -83,12 +81,14 @@ pub struct MetropolisHastings {
 }
 
 impl MetropolisHastings {
-    /// `spread` should be between 0 and 1 both not included and number_of_update should be greater
+    /// `spread` should be between 0 and 1 both not included and `number_of_update` should be greater
     /// than 0. `0.1_f64` is a good choice for this parameter.
     ///
     /// `number_of_update` is the number of times a link matrix is randomly changed.
     /// `spread` is the spread factor for the random matrix change
     /// ( used in [`su3::random_su3_close_to_unity`]).
+    #[inline]
+    #[must_use]
     pub fn new(number_of_update: usize, spread: Real) -> Option<Self> {
         if number_of_update == 0 || !(spread > 0_f64 && spread < 1_f64) {
             return None;
@@ -101,23 +101,29 @@ impl MetropolisHastings {
 
     getter_copy!(
         /// Get the number of attempted updates per steps.
+        #[inline]
+        #[must_use]
         pub const number_of_update() -> usize
     );
 
     getter_copy!(
         /// Get the spread parameter.
+        #[inline]
+        #[must_use]
         pub const spread() -> Real
     );
 }
 
 impl Default for MetropolisHastings {
+    #[inline]
     fn default() -> Self {
-        Self::new(1, 0.1_f64).unwrap()
+        Self::new(1, 0.1_f64).expect("always exist")
     }
 }
 
-impl std::fmt::Display for MetropolisHastings {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+impl Display for MetropolisHastings {
+    #[inline]
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(
             f,
             "Metropolis-Hastings method with {} update and spread {}",
@@ -133,6 +139,7 @@ where
 {
     type Error = State::Error;
 
+    #[inline]
     fn potential_next_element<Rng>(
         &mut self,
         state: &State,
@@ -141,7 +148,7 @@ where
     where
         Rng: rand::Rng + ?Sized,
     {
-        let d = rand::distributions::Uniform::new(0, state.link_matrix().len());
+        let d = Uniform::new(0, state.link_matrix().len());
         let mut link_matrix = state.link_matrix().data().clone();
         (0..self.number_of_update).for_each(|_| {
             let pos = d.sample(rng);
@@ -174,12 +181,14 @@ pub struct MetropolisHastingsDiagnostic {
 }
 
 impl MetropolisHastingsDiagnostic {
-    /// `spread` should be between 0 and 1 both not included and number_of_update should be greater
+    /// `spread` should be between 0 and 1 both not included and `number_of_update` should be greater
     /// than 0. `0.1_f64` is a good choice for this parameter.
     ///
     /// `number_of_update` is the number of times a link matrix is randomly changed.
     /// `spread` is the spread factor for the random matrix change
     /// ( used in [`su3::random_su3_close_to_unity`]).
+    #[inline]
+    #[must_use]
     pub fn new(number_of_update: usize, spread: Real) -> Option<Self> {
         if number_of_update == 0 || spread <= 0_f64 || spread >= 1_f64 {
             return None;
@@ -193,34 +202,44 @@ impl MetropolisHastingsDiagnostic {
     }
 
     /// Get the last probably of acceptance of the random change.
+    #[inline]
+    #[must_use]
     pub const fn prob_replace_last(&self) -> Real {
         self.prob_replace_last
     }
 
     /// Get if last step has accepted the replacement.
+    #[inline]
+    #[must_use]
     pub const fn has_replace_last(&self) -> bool {
         self.has_replace_last
     }
 
     getter_copy!(
         /// Get the number of updates per steps.
+        #[inline]
+        #[must_use]
         pub const number_of_update() -> usize
     );
 
     getter_copy!(
         /// Get the spread parameter.
+        #[inline]
+        #[must_use]
         pub const spread() -> Real
     );
 }
 
 impl Default for MetropolisHastingsDiagnostic {
+    #[inline]
     fn default() -> Self {
-        Self::new(1, 0.1_f64).unwrap()
+        Self::new(1, 0.1_f64).expect("always exist")
     }
 }
 
-impl std::fmt::Display for MetropolisHastingsDiagnostic {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+impl Display for MetropolisHastingsDiagnostic {
+    #[inline]
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(
             f,
             "Metropolis-Hastings method with {} update and spread {}, with diagnostics: has accepted last step {}, probability of acceptance of last step {}",
@@ -238,6 +257,7 @@ where
 {
     type Error = State::Error;
 
+    #[inline]
     fn potential_next_element<Rng>(
         &mut self,
         state: &State,
@@ -246,7 +266,7 @@ where
     where
         Rng: rand::Rng + ?Sized,
     {
-        let d = rand::distributions::Uniform::new(0, state.link_matrix().len());
+        let d = Uniform::new(0, state.link_matrix().len());
         let mut link_matrix = state.link_matrix().data().clone();
         (0..self.number_of_update).for_each(|_| {
             let pos = d.sample(rng);
@@ -259,6 +279,8 @@ where
         )
     }
 
+    #[allow(clippy::unwrap_in_result)] // expect does not panic because of the clamp
+    #[inline]
     fn next_element_default<Rng>(
         &mut self,
         state: State,
@@ -268,16 +290,13 @@ where
         Rng: rand::Rng + ?Sized,
     {
         let potential_next = self.potential_next_element(&state, rng)?;
-        let proba = Self::probability_of_replacement(&state, &potential_next)
-            .min(1_f64)
-            .max(0_f64);
+        let proba = Self::probability_of_replacement(&state, &potential_next).clamp(0_f64, 1_f64); // cspell: ignore proba
         self.prob_replace_last = proba;
-        let d = rand::distributions::Bernoulli::new(proba).unwrap();
+        let d = Bernoulli::new(proba).expect("always exist because of the clamp");
         if d.sample(rng) {
             self.has_replace_last = true;
             Ok(potential_next)
-        }
-        else {
+        } else {
             self.has_replace_last = false;
             Ok(state)
         }
@@ -303,6 +322,8 @@ pub struct MetropolisHastingsDeltaDiagnostic<Rng: rand::Rng> {
 impl<Rng: rand::Rng> MetropolisHastingsDeltaDiagnostic<Rng> {
     getter_copy!(
         /// Get the last probably of acceptance of the random change.
+        #[inline]
+        #[must_use]
         pub const,
         prob_replace_last,
         Real
@@ -310,6 +331,8 @@ impl<Rng: rand::Rng> MetropolisHastingsDeltaDiagnostic<Rng> {
 
     getter_copy!(
         /// Get if last step has accepted the replacement.
+        #[inline]
+        #[must_use]
         pub const,
         has_replace_last,
         bool
@@ -317,6 +340,8 @@ impl<Rng: rand::Rng> MetropolisHastingsDeltaDiagnostic<Rng> {
 
     getter!(
         /// Get a ref to the rng.
+        #[inline]
+        #[must_use]
         pub const,
         rng,
         Rng
@@ -324,20 +349,26 @@ impl<Rng: rand::Rng> MetropolisHastingsDeltaDiagnostic<Rng> {
 
     getter_copy!(
         /// Get the spread parameter.
+        #[inline]
+        #[must_use]
         pub const spread() -> Real
     );
 
     /// Get a mutable reference to the rng.
+    #[inline]
+    #[must_use]
     pub fn rng_mut(&mut self) -> &mut Rng {
         &mut self.rng
     }
 
-    /// `spread` should be between 0 and 1 both not included and number_of_update should be greater
+    /// `spread` should be between 0 and 1 both not included and `number_of_update` should be greater
     /// than 0.
     ///
     /// `number_of_update` is the number of times a link matrix is randomly changed.
     /// `spread` is the spread factor for the random matrix change
     /// ( used in [`su3::random_su3_close_to_unity`]).
+    #[inline]
+    #[must_use]
     pub fn new(spread: Real, rng: Rng) -> Option<Self> {
         if spread <= 0_f64 || spread >= 1_f64 {
             return None;
@@ -351,36 +382,40 @@ impl<Rng: rand::Rng> MetropolisHastingsDeltaDiagnostic<Rng> {
     }
 
     /// Absorbs self and return the RNG as owned. It essentially deconstruct the structure.
-    #[allow(clippy::missing_const_for_fn)] // false positive
+    #[inline]
+    #[must_use]
     pub fn rng_owned(self) -> Rng {
         self.rng
     }
 
     #[inline]
+    #[must_use]
     fn delta_s<const D: usize>(
         link_matrix: &LinkMatrix,
         lattice: &LatticeCyclic<D>,
         link: &LatticeLinkCanonical<D>,
-        new_link: &na::Matrix3<Complex>,
+        new_link: &nalgebra::Matrix3<Complex>,
         beta: Real,
     ) -> Real {
         let old_matrix = link_matrix
             .matrix(&LatticeLink::from(*link), lattice)
-            .unwrap();
+            .expect("matrix not found");
         delta_s_old_new_cmp(link_matrix, lattice, link, new_link, beta, &old_matrix)
     }
 
+    // cspell: ignore modif
     #[inline]
+    #[must_use]
     fn potential_modif<const D: usize>(
         &mut self,
         state: &LatticeStateDefault<D>,
-    ) -> (LatticeLinkCanonical<D>, na::Matrix3<Complex>) {
-        let d_p = rand::distributions::Uniform::new(0, state.lattice().dim());
-        let d_d = rand::distributions::Uniform::new(0, LatticeCyclic::<D>::dim_st());
+    ) -> (LatticeLinkCanonical<D>, nalgebra::Matrix3<Complex>) {
+        let d_p = Uniform::new(0, state.lattice().dim());
+        let d_d = Uniform::new(0, LatticeCyclic::<D>::dim_st());
 
         let point = LatticePoint::from_fn(|_| d_p.sample(&mut self.rng));
         let direction = Direction::positive_directions()[d_d.sample(&mut self.rng)];
-        let link = LatticeLinkCanonical::new(point, direction).unwrap();
+        let link = LatticeLinkCanonical::new(point, direction).expect("unreachable");
         let index = link.to_index(state.lattice());
 
         let old_link_m = state.link_matrix()[index];
@@ -391,6 +426,7 @@ impl<Rng: rand::Rng> MetropolisHastingsDeltaDiagnostic<Rng> {
     }
 
     #[inline]
+    #[must_use]
     fn next_element_default<const D: usize>(
         &mut self,
         mut state: LatticeStateDefault<D>,
@@ -403,14 +439,13 @@ impl<Rng: rand::Rng> MetropolisHastingsDeltaDiagnostic<Rng> {
             &matrix,
             state.beta(),
         );
-        let proba = (-delta_s).exp().min(1_f64).max(0_f64);
+        let proba = (-delta_s).exp().clamp(0_f64, 1_f64);
         self.prob_replace_last = proba;
-        let d = rand::distributions::Bernoulli::new(proba).unwrap();
+        let d = Bernoulli::new(proba).expect("always exist");
         if d.sample(&mut self.rng) {
             self.has_replace_last = true;
-            *state.link_mut(&link).unwrap() = matrix;
-        }
-        else {
+            *state.link_mut(&link).expect("link not found") = matrix;
+        } else {
             self.has_replace_last = false;
         }
         state
@@ -418,15 +453,15 @@ impl<Rng: rand::Rng> MetropolisHastingsDeltaDiagnostic<Rng> {
 }
 
 impl<Rng: rand::Rng + Default> Default for MetropolisHastingsDeltaDiagnostic<Rng> {
+    #[inline]
     fn default() -> Self {
-        Self::new(0.1_f64, Rng::default()).unwrap()
+        Self::new(0.1_f64, Rng::default()).expect("unreachable")
     }
 }
 
-impl<Rng: rand::Rng + std::fmt::Display> std::fmt::Display
-    for MetropolisHastingsDeltaDiagnostic<Rng>
-{
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+impl<Rng: rand::Rng + Display> Display for MetropolisHastingsDeltaDiagnostic<Rng> {
+    #[inline]
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(
             f,
             "Metropolis-Hastings delta method with rng {} and spread {}, with diagnostics: has accepted last step {}, probability of acceptance of last step {}",
@@ -439,12 +474,14 @@ impl<Rng: rand::Rng + std::fmt::Display> std::fmt::Display
 }
 
 impl<Rng: rand::Rng> AsRef<Rng> for MetropolisHastingsDeltaDiagnostic<Rng> {
+    #[inline]
     fn as_ref(&self) -> &Rng {
         self.rng()
     }
 }
 
 impl<Rng: rand::Rng> AsMut<Rng> for MetropolisHastingsDeltaDiagnostic<Rng> {
+    #[inline]
     fn as_mut(&mut self) -> &mut Rng {
         self.rng_mut()
     }
@@ -469,29 +506,34 @@ where
 #[cfg(test)]
 mod test {
 
-    use rand::SeedableRng;
+    use std::error::Error;
+
+    use rand::{rngs::StdRng, SeedableRng};
 
     use super::*;
-    use crate::simulation::state::*;
+    use crate::{error::ImplementationError, simulation::state::*};
 
     const SEED: u64 = 0x45_78_93_f4_4a_b0_67_f0;
 
     #[test]
-    fn test_mh_delta() {
-        let mut rng = rand::rngs::StdRng::seed_from_u64(SEED);
+    fn test_mh_delta() -> Result<(), Box<dyn Error>> {
+        let mut rng = StdRng::seed_from_u64(SEED);
 
         let size = 1_000_f64;
         let number_of_pts = 4;
         let beta = 2_f64;
         let mut simulation =
-            LatticeStateDefault::<4>::new_determinist(size, beta, number_of_pts, &mut rng).unwrap();
+            LatticeStateDefault::<4>::new_determinist(size, beta, number_of_pts, &mut rng)?;
 
-        let mut mcd = MetropolisHastingsDeltaDiagnostic::new(0.01_f64, rng).unwrap();
+        let mut mcd = MetropolisHastingsDeltaDiagnostic::new(0.01_f64, rng)
+            .ok_or(ImplementationError::OptionWithUnexpectedNone)?;
         for _ in 0_u32..10_u32 {
             let mut simulation2 = simulation.clone();
             let (link, matrix) = mcd.potential_modif(&simulation);
-            *simulation2.link_mut(&link).unwrap() = matrix;
-            let ds = MetropolisHastingsDeltaDiagnostic::<rand::rngs::StdRng>::delta_s(
+            *simulation2
+                .link_mut(&link)
+                .ok_or(ImplementationError::OptionWithUnexpectedNone)? = matrix;
+            let ds = MetropolisHastingsDeltaDiagnostic::<StdRng>::delta_s(
                 simulation.link_matrix(),
                 simulation.lattice(),
                 &link,
@@ -506,38 +548,42 @@ mod test {
             let prob_of_replacement = (simulation.hamiltonian_links()
                 - simulation2.hamiltonian_links())
             .exp()
-            .min(1_f64)
-            .max(0_f64);
-            assert!(((-ds).exp().min(1_f64).max(0_f64) - prob_of_replacement).abs() < 1E-8_f64);
+            .clamp(0_f64, 1_f64);
+            assert!(((-ds).exp().clamp(0_f64, 1_f64) - prob_of_replacement).abs() < 1E-8_f64);
             simulation = simulation2;
         }
+        Ok(())
     }
     #[test]
-    fn methods_common_traits() {
+    fn methods_common_traits() -> Result<(), ImplementationError> {
         assert_eq!(
-            MetropolisHastings::default(),
-            MetropolisHastings::new(1, 0.1_f64).unwrap()
+            Some(MetropolisHastings::default()),
+            MetropolisHastings::new(1, 0.1_f64)
         );
         assert_eq!(
-            MetropolisHastingsDiagnostic::default(),
-            MetropolisHastingsDiagnostic::new(1, 0.1_f64).unwrap()
+            Some(MetropolisHastingsDiagnostic::default()),
+            MetropolisHastingsDiagnostic::new(1, 0.1_f64)
         );
 
-        let rng = rand::rngs::StdRng::seed_from_u64(SEED);
+        let rng = StdRng::seed_from_u64(SEED);
         assert!(MetropolisHastingsDeltaDiagnostic::new(0_f64, rng.clone()).is_none());
         assert!(MetropolisHastings::new(0, 0.1_f64).is_none());
         assert!(MetropolisHastingsDiagnostic::new(1, 0_f64).is_none());
 
         assert_eq!(
-            MetropolisHastings::new(2, 0.2_f64).unwrap().to_string(),
+            MetropolisHastings::new(2, 0.2_f64)
+                .ok_or(ImplementationError::OptionWithUnexpectedNone)?
+                .to_string(),
             "Metropolis-Hastings method with 2 update and spread 0.2"
         );
         assert_eq!(
-            MetropolisHastingsDiagnostic::new(2, 0.2_f64).unwrap().to_string(),
+            MetropolisHastingsDiagnostic::new(2, 0.2_f64).ok_or(ImplementationError::OptionWithUnexpectedNone)?.to_string(),
             "Metropolis-Hastings method with 2 update and spread 0.2, with diagnostics: has accepted last step false, probability of acceptance of last step 0"
         );
-        let mut mhdd = MetropolisHastingsDeltaDiagnostic::new(0.1_f64, rng).unwrap();
-        let _: &rand::rngs::StdRng = mhdd.as_ref();
-        let _: &mut rand::rngs::StdRng = mhdd.as_mut();
+        let mut mhdd = MetropolisHastingsDeltaDiagnostic::new(0.1_f64, rng)
+            .ok_or(ImplementationError::OptionWithUnexpectedNone)?; // cspell: ignore mhdd
+        let _: &StdRng = mhdd.as_ref();
+        let _: &mut StdRng = mhdd.as_mut();
+        Ok(())
     }
 }
